backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
python-multipart>=0.0.9
torch>=2.2.0
transformers>=4.57.0
sentencepiece>=0.2.0
Pillow>=10.0.0
PyYAML>=6.0.1
safetensors>=0.4.3
